@@ -4,7 +4,7 @@
   "version": "1.0.0",
   "author": "Wes Bos <wesbos@gmail.com>",
   "engines": {
-    "node": ">=12.14.0"
+    "node": ">=16.13.1"
   },
   "dependencies": {
     "@mdx-js/loader": "2.1.5",
