@@ -1,4 +1,6 @@
 const chrome = require('chrome-aws-lambda');
+// Switch to this, re: https://answers.netlify.com/t/netlify-function-with-puppeteer-breaks-if-i-make-any-changes/76924/8
+// const chrome = require('@sparticuz/chromium');
 const puppeteer = require('puppeteer-core');
 const wait = require('waait');
 
@@ -43,17 +45,16 @@ async function getScreenshot(url, isDev) {
 }
 
 // Docs on event and context https://www.netlify.com/docs/functions/#the-handler-method
-exports.handler = async (event, context) => {
+exports.handler = async (event) => {
   const qs = new URLSearchParams(event.queryStringParameters);
-  console.log(
-    `${process.env.URL || `http://localhost:8888`}/thumbnail?${qs.toString()}`
-  );
+  console.log(`${process.env.URL || `http://localhost:8888`}/thumbnail?${qs.toString()}`);
   const photoBuffer = await getScreenshot(
     `${process.env.URL || `http://localhost:8888`}/thumbnail?${qs.toString()}`,
     // Here we need to pass a boolean to say if we are on the server. Netlify has a bug where process.env.NETLIFY is undefiend in functions so I'm using one of the only vars I can find
     // !process.env.NETLIFY
     process.env.URL.includes('http://localhost')
   );
+  console.log('x');
   return {
     statusCode: 200,
     body: photoBuffer,
