@@ -1,3 +1,4 @@
+/* eslint-disable */
 import fetch from 'isomorphic-fetch';
 import { promises as fs } from 'fs';
 import { Html5Entities } from 'html-entities';
@@ -5,15 +6,16 @@ import replaceAll from 'string.prototype.replaceall';
 import jsdom, { JSDOM } from 'jsdom';
 import FileType from 'file-type';
 import getSlug from 'speakingurl';
-import tips from './tips';
 import urlExpander from 'expand-url';
 import { promisify } from 'util';
+import tips from './tips';
+
 const expand = promisify(urlExpander.expand);
 
 const DLIMAGES = false;
 
 // const expression = /[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)?/gi;
-const expression = /(?:(?:https?|ftp|file):\/\/|www\.|ftp\.)(?:\([-A-Z0-9+&@#\/%=~_|$?!:,.]*\)|[-A-Z0-9+&@#\/%=~_|$?!:,.])*(?:\([-A-Z0-9+&@#\/%=~_|$?!:,.]*\)|[A-Z0-9+&@#\/%=~_|$])/igm;
+const expression = /(?:(?:https?|ftp|file):\/\/|www\.|ftp\.)(?:\([-A-Z0-9+&@#\/%=~_|$?!:,.]*\)|[-A-Z0-9+&@#\/%=~_|$?!:,.])*(?:\([-A-Z0-9+&@#\/%=~_|$?!:,.]*\)|[A-Z0-9+&@#\/%=~_|$])/gim;
 const linkRegex = new RegExp(expression);
 
 function getImageName(path) {
@@ -24,37 +26,23 @@ function findImagePaths(content) {
   const virtualConsole = new jsdom.VirtualConsole();
   // virtualConsole.on('log', console.log);
   const DOM = new JSDOM(content);
-  const imgs = DOM.window.document.querySelectorAll(
-    '.AdaptiveMedia-container img'
-  );
-  const videos = DOM.window.document.querySelectorAll(
-    '[data-playable-media-url]'
-  );
+  const imgs = DOM.window.document.querySelectorAll('.AdaptiveMedia-container img');
+  const videos = DOM.window.document.querySelectorAll('[data-playable-media-url]');
   return {
-    images: Array.from(imgs).map(img => img.src),
-    videos: Array.from(videos).map(video =>
-      video.style['background-image']
-        .replace('url(', '')
-        .replace(')', '')
-        .replace('.jpg', '.mp4')
-        .replace('tweet_video_thumb', 'tweet_video')
-    ),
+    images: Array.from(imgs).map((img) => img.src),
+    videos: Array.from(videos).map((video) => video.style['background-image'].replace('url(', '').replace(')', '').replace('.jpg', '.mp4').replace('tweet_video_thumb', 'tweet_video')),
     // https://video.twimg.com/tweet_video/EIodxyoXYAI8fLE.mp4
   };
 }
 
-
 async function downloadImage(remotePath, localFolder) {
   console.log(`Downloading ${remotePath} to ${localFolder}`);
-  const imageData = await fetch(remotePath).then(res => res.buffer());
+  const imageData = await fetch(remotePath).then((res) => res.buffer());
   console.log(`~~~Doing ${remotePath}`);
-  const { ext = 'png' } = await FileType.fromBuffer(imageData) || {};
+  const { ext = 'png' } = (await FileType.fromBuffer(imageData)) || {};
   const imageName = getImageName(remotePath);
   const [, extension] = imageName.split('.');
-  await fs.writeFile(
-    `${localFolder}/${imageName}${extension ? '' : `.${ext}`}`,
-    imageData
-  );
+  await fs.writeFile(`${localFolder}/${imageName}${extension ? '' : `.${ext}`}`, imageData);
 }
 
 // const selectedTips = tips.slice(35, 40);
@@ -69,34 +57,31 @@ const selectedTips = tips.slice(85);
 async function getTweets() {
   await fs.mkdir(`./src/tips/`, { recursive: true });
   for (const tip of selectedTips) {
-    const slug = getSlug(
-      tip.text
-        .split(' ')
-        .slice(0, 8)
-        .join(' ')
-    );
+    const slug = getSlug(tip.text.split(' ').slice(0, 8).join(' '));
     const folderPath = `./src/tips/${slug}`;
     await fs.mkdir(folderPath, { recursive: true });
-    const html = await fetch(tip.url).then(x => x.text());
+    const html = await fetch(tip.url).then((x) => x.text());
     const { images, videos } = await findImagePaths(html);
 
     // find links
     const links = tip.text.match(linkRegex) || [];
     const resourceLinks = links.slice(0, links.length - 1);
     console.log(resourceLinks);
-    const expandedLinks = await Promise.all(resourceLinks.map(link => {
-      const url = link.startsWith('http') ? link : `http://${link}`;
-      return expand(url).catch(console.error);
-    }));
+    const expandedLinks = await Promise.all(
+      resourceLinks.map((link) => {
+        const url = link.startsWith('http') ? link : `http://${link}`;
+        return expand(url).catch(console.error);
+      })
+    );
 
     console.log(expandedLinks);
     // Download images and videos
     console.log('finding image paths');
     console.log(videos);
-    await Promise.all(images.map(path => downloadImage(path, folderPath)));
-    await Promise.all(videos.map(path => downloadImage(path, folderPath)));
+    await Promise.all(images.map((path) => downloadImage(path, folderPath)));
+    await Promise.all(videos.map((path) => downloadImage(path, folderPath)));
 
-    links.forEach(link => {
+    links.forEach((link) => {
       tip.text = tip.text.replace(link, '');
       tip.text = tip.text.replace('http://', '');
       tip.text = tip.text.replace('http://', '');
@@ -111,11 +96,11 @@ tweetURL: ${tip.url}
 text: ${tip.text}
 slug: ${slug}
 images:
-  ${images.map(x => `- ${x}`).join('\n')}
+  ${images.map((x) => `- ${x}`).join('\n')}
 videos:
-  ${videos.map(x => `- ${x}`).join('\n')}
+  ${videos.map((x) => `- ${x}`).join('\n')}
 links:
-  ${expandedLinks.map(x => `- ${x}`).join('\n')}
+  ${expandedLinks.map((x) => `- ${x}`).join('\n')}
 ---
 ${tip.text.trim()}
 `;
@@ -129,4 +114,3 @@ ${tip.text.trim()}
 }
 
 getTweets();
-
