@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 
 export default function LatestCourse() {
   return (
@@ -6,20 +6,9 @@ export default function LatestCourse() {
       <h3>
         <span className="highlight">Master Gatsby</span>
       </h3>
-      <img
-        src="https://res.cloudinary.com/wesbos/image/fetch/w_700,q_auto,f_auto/https://courses.wesbos.com/images/GAT/GAT-social-share.png"
-        alt="Master Gatsby"
-      />
-      <p>
-        Building modern websites is tough. Preloading, routing, compression,
-        critical CSS, caching, scaling and bundlers all make for blazing fast
-        websites, but extra development and tooling get in the way.
-      </p>
-      <p>
-        Gatsby is a React.js framework that does it all for you. This course
-        will teach you how to build your websites and let Gatsby take care of
-        all the Hard Stuff™.
-      </p>
+      <img src="https://res.cloudinary.com/wesbos/image/fetch/w_700,q_auto,f_auto/https://courses.wesbos.com/images/GAT/GAT-social-share.png" alt="Master Gatsby" />
+      <p>Building modern websites is tough. Preloading, routing, compression, critical CSS, caching, scaling and bundlers all make for blazing fast websites, but extra development and tooling get in the way.</p>
+      <p>Gatsby is a React.js framework that does it all for you. This course will teach you how to build your websites and let Gatsby take care of all the Hard Stuff™.</p>
       <a href="https://MasterGatsby.com">MasterGatsby.com</a>
     </div>
   );
