@@ -47,29 +47,19 @@ const PaginationStyles = styled.div`
   }
 `;
 
-export default function Pagination({
-  totalCount,
-  currentPage = 1,
-  pathPrefix,
-}) {
+export default function Pagination({ totalCount, currentPage = 1, pathPrefix }) {
   const totalPages = Math.ceil(totalCount / 10);
   const nextPage = currentPage + 1;
   const prevPage = currentPage - 1;
   return (
     <PaginationStyles>
-      <Link
-        disabled={prevPage <= 0 ? true : null}
-        to={`${pathPrefix}${prevPage}`}
-      >
+      <Link disabled={prevPage <= 0 ? true : null} to={`${pathPrefix}${prevPage}`}>
         ← Prev 10 please
       </Link>
       <p>
         Page {currentPage} of {totalPages}
       </p>
-      <Link
-        disabled={nextPage > totalPages ? true : null}
-        to={nextPage > totalPages ? null : `${pathPrefix}${nextPage}`}
-      >
+      <Link disabled={nextPage > totalPages ? true : null} to={nextPage > totalPages ? null : `${pathPrefix}${nextPage}`}>
         {nextPage > totalPages ? `That's all Folks` : `10 More please →`}
       </Link>
     </PaginationStyles>
