@@ -1,6 +1,5 @@
-import styled from 'styled-components'
+import styled from 'styled-components';
 
-const Pre = styled.pre`
-`;
+const Pre = styled.pre``;
 
 export default Pre;
