@@ -11,50 +11,38 @@ const Blog = function ({ data, pageContext, path }) {
   if (!data) return <p>Shooooot! No Post found!</p>;
   return (
     <>
-      <Pagination
-        currentPage={pageContext.currentPage}
-        totalCount={data.allMdx.totalCount}
-        pathPrefix="/blog/"
-      />
+      <Pagination currentPage={pageContext.currentPage} totalCount={data.allMdx.totalCount} pathPrefix="/blog/" />
       <PostMetaTags
         post={{
           frontmatter: {
             slug: path,
-            title: `Blog ${
-              pageContext.currentPage ? `- Page ${pageContext.currentPage}` : ''
-            }`,
+            title: `Blog ${pageContext.currentPage ? `- Page ${pageContext.currentPage}` : ''}`,
           },
         }}
       />
       <PostGrid>
         {data.allMdx &&
-          data.allMdx.edges.map(function ({ node: post }) {
+          data.allMdx.edges.map(({ node: post }) => {
             const hasImage = !!post.frontmatter.image;
             return (
               <PostGridItem key={post.id} hasImage={hasImage}>
-                {post.frontmatter.image &&
-                  post.frontmatter.image.childImageSharp && (
-                    <div>
-                      {hasImage && (
-                        <Link to={post.fields.slug}>
-                          <Img image={post.frontmatter.image} />
-                        </Link>
-                      )}
-                      <div className="postMeta">
-                        <time dateTime={post.frontmatter.date}>
-                          {format(
-                            new Date(post.frontmatter.date),
-                            'MMMM d, yyyy'
-                          )}
-                        </time>
-                        <ul className="categories">
-                          {post.frontmatter.category.map((cat) => (
-                            <li key={cat}>{cat}</li>
-                          ))}
-                        </ul>
-                      </div>
+                {post.frontmatter.image && post.frontmatter.image.childImageSharp && (
+                  <div>
+                    {hasImage && (
+                      <Link to={post.fields.slug}>
+                        <Img image={post.frontmatter.image} />
+                      </Link>
+                    )}
+                    <div className="postMeta">
+                      <time dateTime={post.frontmatter.date}>{format(new Date(post.frontmatter.date), 'MMMM d, yyyy')}</time>
+                      <ul className="categories">
+                        {post.frontmatter.category.map((cat) => (
+                          <li key={cat}>{cat}</li>
+                        ))}
+                      </ul>
                     </div>
-                  )}
+                  </div>
+                )}
                 <div>
                   <H as="h3">
                     <Link to={post.fields.slug}>{post.frontmatter.title}</Link>
@@ -65,11 +53,7 @@ const Blog = function ({ data, pageContext, path }) {
             );
           })}
       </PostGrid>
-      <Pagination
-        currentPage={pageContext.currentPage}
-        totalCount={data.allMdx.totalCount}
-        pathPrefix="/blog/"
-      />
+      <Pagination currentPage={pageContext.currentPage} totalCount={data.allMdx.totalCount} pathPrefix="/blog/" />
     </>
   );
 };
@@ -84,12 +68,7 @@ export const pageQuery = graphql`
         description
       }
     }
-    allMdx(
-      filter: { fields: { collection: { eq: "post" } } }
-      sort: { fields: [frontmatter___date], order: DESC }
-      limit: 10
-      skip: $skip
-    ) {
+    allMdx(filter: { fields: { collection: { eq: "post" } } }, sort: { fields: [frontmatter___date], order: DESC }, limit: 10, skip: $skip) {
       totalCount
       edges {
         node {
