@@ -1,9 +1,9 @@
 import { Link } from 'gatsby';
-import React, { useState } from 'react';
+import React from 'react';
 import H from '../components/mdxComponents/Headings';
 import { PostMetaTags } from '../components/MetaTags';
 
-export default function AboutPage({ data, path }) {
+export default function AboutPage({ path }) {
   return (
     <>
       <PostMetaTags
@@ -17,267 +17,78 @@ export default function AboutPage({ data, path }) {
       <H>Privacy Policy</H>
       <p>Last updated on January 13, 2022</p>
 
-      <p>
-        This is the Privacy Policy of the of the WesBos.com web site, its
-        related websites, and its and their successor web sites (collectively
-        referred to as the “Site”), which are owned and operated by Bos Type
-        Inc. (which may be referred to in the Terms as “we”, “us”, “Wesbos.com”
-        or “Wes Bos”). It describes the information that we collect from users
-        of our Site (“Users”) as part of the normal operation of our Site, and
-        how we use and disclose this information.
-      </p>
+      <p>This is the Privacy Policy of the of the WesBos.com web site, its related websites, and its and their successor web sites (collectively referred to as the “Site”), which are owned and operated by Bos Type Inc. (which may be referred to in the Terms as “we”, “us”, “Wesbos.com” or “Wes Bos”). It describes the information that we collect from users of our Site (“Users”) as part of the normal operation of our Site, and how we use and disclose this information.</p>
 
       <p>
-        By accepting the Privacy Policy in registration or by visiting and using
-        the Site, you expressly consent to our collection, use and disclosure of
-        your information in accordance with this Privacy Policy. This Privacy
-        Policy is incorporated into and subject to our Terms and Conditions of
-        Use <Link to="/terms">(the “Terms”)</Link>.
+        By accepting the Privacy Policy in registration or by visiting and using the Site, you expressly consent to our collection, use and disclosure of your information in accordance with this Privacy Policy. This Privacy Policy is incorporated into and subject to our Terms and Conditions of Use <Link to="/terms">(the “Terms”)</Link>.
       </p>
 
       <H as="h2">How We Use Your Information</H>
 
-      <p>
-        We use information in the file we maintain about you, and other
-        information we obtain from your current and past activities on the Site,
-        to provide to you the services offered by the Site; resolve service
-        disputes; troubleshoot problems; measure consumer interest in our
-        products and services, inform you about online and offline offers,
-        products, services, events and updates; deliver information to you that,
-        in some cases, is relevant to your interests, such as product news;
-        customize your experience; detect and protect us against error, fraud
-        and other criminal activity; enforce the Terms; provide you with system
-        or administrative messages, and as otherwise described to you at the
-        time of collection.
-      </p>
+      <p>We use information in the file we maintain about you, and other information we obtain from your current and past activities on the Site, to provide to you the services offered by the Site; resolve service disputes; troubleshoot problems; measure consumer interest in our products and services, inform you about online and offline offers, products, services, events and updates; deliver information to you that, in some cases, is relevant to your interests, such as product news; customize your experience; detect and protect us against error, fraud and other criminal activity; enforce the Terms; provide you with system or administrative messages, and as otherwise described to you at the time of collection.</p>
 
-      <p>
-        We may also use information about you to improve our marketing and
-        promotional efforts, to analyze Site usage, to improve our content and
-        product offerings, and to customize the Site's content, layout, and
-        services. These uses improve the Site and better tailor it to meet your
-        needs, so as to provide you with a smooth, efficient, safe and
-        customized experience while using the Site.
-      </p>
+      <p>We may also use information about you to improve our marketing and promotional efforts, to analyze Site usage, to improve our content and product offerings, and to customize the Site's content, layout, and services. These uses improve the Site and better tailor it to meet your needs, so as to provide you with a smooth, efficient, safe and customized experience while using the Site.</p>
 
       <H as="h2">What Information Do We Collect?</H>
 
-      <p>
-        Our primary purpose in collecting information from you is to provide you
-        with a safe, smooth, efficient, and customized experience. This allows
-        us to provide services and features that most likely meet your needs,
-        and to customize our service to make your experience safer and easier.
-        We only collect information about you that we consider necessary for
-        achieving this purpose. We do not collect any credit card or other
-        permitted payment method information from you – all payment processing
-        is done by our third party payment processors.
-      </p>
+      <p>Our primary purpose in collecting information from you is to provide you with a safe, smooth, efficient, and customized experience. This allows us to provide services and features that most likely meet your needs, and to customize our service to make your experience safer and easier. We only collect information about you that we consider necessary for achieving this purpose. We do not collect any credit card or other permitted payment method information from you – all payment processing is done by our third party payment processors.</p>
 
-      <p>
-        In general, you can browse the Site without telling us who you are or
-        revealing any personal information about yourself. Once you become a
-        registered User or begin to make a purchase from the Site, we require
-        you to provide various contact and identity information and other
-        information as indicated on the relevant forms on the Site, and you are
-        no longer anonymous to us. Where possible, on these forms we indicate
-        which fields are required and which fields are optional. In addition, as
-        you use the Site, you can from time to time enter or send to us personal
-        information.
-      </p>
+      <p>In general, you can browse the Site without telling us who you are or revealing any personal information about yourself. Once you become a registered User or begin to make a purchase from the Site, we require you to provide various contact and identity information and other information as indicated on the relevant forms on the Site, and you are no longer anonymous to us. Where possible, on these forms we indicate which fields are required and which fields are optional. In addition, as you use the Site, you can from time to time enter or send to us personal information.</p>
 
-      <p>
-        You always have the option to not provide information by choosing not to
-        become a User or by not using the particular feature of the Site for
-        which the information is being collected.
-      </p>
+      <p>You always have the option to not provide information by choosing not to become a User or by not using the particular feature of the Site for which the information is being collected.</p>
 
-      <p>
-        We also automatically track certain information about you based upon
-        your behaviour on our Site. We use this information to do internal
-        research on our users' demographics, interests, and behaviour to better
-        understand, protect and serve you and our community. This information
-        may include the URL that you just came from (whether this URL is on our
-        Site or not), which URL you next go to (whether this URL is on our Site
-        or not), your computer browser information, your IP address and your
-        location.
-      </p>
+      <p>We also automatically track certain information about you based upon your behaviour on our Site. We use this information to do internal research on our users' demographics, interests, and behaviour to better understand, protect and serve you and our community. This information may include the URL that you just came from (whether this URL is on our Site or not), which URL you next go to (whether this URL is on our Site or not), your computer browser information, your IP address and your location.</p>
 
-      <p>
-        “Cookies” are small files placed on your hard drive that assist us in
-        providing our services. We use cookies to allow you to enter your
-        password less frequently during a session, and we use data collection
-        devices (such as Google Analytics), including cookies, on certain pages
-        of the Site to help analyze our web page flow, measure promotional
-        effectiveness, and promote trust and safety, to offer certain features
-        that are only available through the use of a cookie and to allow us to
-        provide information that is targeted to your interests.
-      </p>
+      <p>“Cookies” are small files placed on your hard drive that assist us in providing our services. We use cookies to allow you to enter your password less frequently during a session, and we use data collection devices (such as Google Analytics), including cookies, on certain pages of the Site to help analyze our web page flow, measure promotional effectiveness, and promote trust and safety, to offer certain features that are only available through the use of a cookie and to allow us to provide information that is targeted to your interests.</p>
 
-      <p>
-        Following your registration on the Site, you can review and change your
-        information in your account area. If you disclose personal information
-        on the Site and wish to have it removed, please contact us at the
-        support contact information posted on the Site.
-      </p>
+      <p>Following your registration on the Site, you can review and change your information in your account area. If you disclose personal information on the Site and wish to have it removed, please contact us at the support contact information posted on the Site.</p>
 
       <H as="h2">Our Disclosure of Your Information</H>
 
-      <p>
-        We will not sell or rent your personally identifiable information to
-        third parties without your explicit consent. The following describes
-        some of the ways that your information may be disclosed in the normal
-        scope of business to provide our services:
-      </p>
+      <p>We will not sell or rent your personally identifiable information to third parties without your explicit consent. The following describes some of the ways that your information may be disclosed in the normal scope of business to provide our services:</p>
 
-      <p>
-        Anonymized Aggregated Data. We aggregate and anonymize data and use and
-        disclose such information for a variety of purposes. However, in these
-        situations, we do not disclose any information that could be used to
-        identify you personally.
-      </p>
+      <p>Anonymized Aggregated Data. We aggregate and anonymize data and use and disclose such information for a variety of purposes. However, in these situations, we do not disclose any information that could be used to identify you personally.</p>
 
-      <p>
-        Public Areas of the Site. All of your activities in the public areas of
-        the Site will be identifiable to your User ID, and other people can see
-        your published content.
-      </p>
+      <p>Public Areas of the Site. All of your activities in the public areas of the Site will be identifiable to your User ID, and other people can see your published content.</p>
 
-      <p>
-        Subsidiaries and Affiliates; Service Providers. We may from time to time
-        use the services of affiliates, subsidiaries and unrelated service
-        providers in the operation of the Site, and may disclose your
-        information to them in the course of our use of their services. For
-        example, we may use the services of third-party hosting companies to
-        host the operation of the Site. This may involve the hosting of data,
-        including your information, on servers operated by those hosting
-        companies. We take care to use only service providers that we believe
-        are reputable and able to live up to our and your expectations,
-        including about the handling of confidential information.
-      </p>
+      <p>Subsidiaries and Affiliates; Service Providers. We may from time to time use the services of affiliates, subsidiaries and unrelated service providers in the operation of the Site, and may disclose your information to them in the course of our use of their services. For example, we may use the services of third-party hosting companies to host the operation of the Site. This may involve the hosting of data, including your information, on servers operated by those hosting companies. We take care to use only service providers that we believe are reputable and able to live up to our and your expectations, including about the handling of confidential information.</p>
 
-      <p>
-        Law and Harm; Jurisdictions. Notwithstanding anything to the contrary in
-        this policy, we may preserve or disclose your information if we believe
-        that it is reasonably necessary to comply with a law, regulation or
-        legal request; to protect the safety of any person; to address fraud,
-        security or technical issues; or to protect our or any other person’s
-        rights or property. However, nothing in this Privacy Policy is intended
-        to limit any legal defenses or objections that you may have to a third
-        party’s, including a government’s, request to disclose your information.
-        Finally, please note that in some cases, information that we collect may
-        be stored or processed outside of Canada. In such cases, we continue to
-        protect the information with appropriate safeguards, but it may be
-        subject to the legal jurisdiction of those countries and governmental
-        authorities in those countries. Please see the bottom of this policy for
-        further information about how to direct any questions you might have
-        about service providers outside of Canada.
-      </p>
+      <p>Law and Harm; Jurisdictions. Notwithstanding anything to the contrary in this policy, we may preserve or disclose your information if we believe that it is reasonably necessary to comply with a law, regulation or legal request; to protect the safety of any person; to address fraud, security or technical issues; or to protect our or any other person’s rights or property. However, nothing in this Privacy Policy is intended to limit any legal defenses or objections that you may have to a third party’s, including a government’s, request to disclose your information. Finally, please note that in some cases, information that we collect may be stored or processed outside of Canada. In such cases, we continue to protect the information with appropriate safeguards, but it may be subject to the legal jurisdiction of those countries and governmental authorities in those countries. Please see the bottom of this policy for further information about how to direct any questions you might have about service providers outside of Canada.</p>
 
-      <p>
-        Sale of Business. We may also disclose your information to the acquiror
-        or its agents in the course of the sale of our business. If we do this,
-        the disclosure will be subject to confidentiality arrangements customary
-        in such transactions.
-      </p>
+      <p>Sale of Business. We may also disclose your information to the acquiror or its agents in the course of the sale of our business. If we do this, the disclosure will be subject to confidentiality arrangements customary in such transactions.</p>
 
       <H as="h2">Security</H>
 
-      <p>
-        We strive to protect your personal information. For security of
-        transactions, we use the Secure Sockets Layer (SSL) protocol, which
-        encrypts any personal information you enter into registration forms on
-        the Site. The encryption process protects your information, by
-        scrambling it before it is sent to us from your computer. We also make
-        commercially reasonable effort to ensure the security of your personal
-        information on our system. Unfortunately, no data transmission over the
-        Internet can be guaranteed to be 100% secure. As a result, while we
-        strive to protect your personal information, we cannot warrant the
-        security of any information you transmit to us.
-      </p>
+      <p>We strive to protect your personal information. For security of transactions, we use the Secure Sockets Layer (SSL) protocol, which encrypts any personal information you enter into registration forms on the Site. The encryption process protects your information, by scrambling it before it is sent to us from your computer. We also make commercially reasonable effort to ensure the security of your personal information on our system. Unfortunately, no data transmission over the Internet can be guaranteed to be 100% secure. As a result, while we strive to protect your personal information, we cannot warrant the security of any information you transmit to us.</p>
 
       <H as="h2">Data Retention</H>
 
-      <p>
-        We will retain your information for as long as your account is active or
-        as needed to provide you services.
-      </p>
+      <p>We will retain your information for as long as your account is active or as needed to provide you services.</p>
 
-      <p>
-        Our goal is to retain your information as necessary to comply with our
-        legal obligations, resolve disputes, and enforce our agreements; this
-        retention period may extend past the point at which you close your
-        account.
-      </p>
+      <p>Our goal is to retain your information as necessary to comply with our legal obligations, resolve disputes, and enforce our agreements; this retention period may extend past the point at which you close your account.</p>
 
-      <p>
-        The criteria that will be for determining periods of retention will be
-        based on the type of data. For example, data relevant to accounting
-        information will be retained for at least 7 years, whereas temporary
-        cookie data of a Site visitor may not be retained at all.
-      </p>
+      <p>The criteria that will be for determining periods of retention will be based on the type of data. For example, data relevant to accounting information will be retained for at least 7 years, whereas temporary cookie data of a Site visitor may not be retained at all.</p>
 
       <H as="h2">Other Information Collectors</H>
 
-      <p>
-        Except as otherwise expressly included in this Privacy Policy, this
-        document only addresses the use and disclosure of information we collect
-        from you. To the extent that you disclose your information to other
-        parties through the Site (including our payment processors), different
-        rules may apply to their use, collection and disclosure of the personal
-        information you disclose to them. Since we do not control the
-        information use, collection or disclosure policies of third parties, you
-        are subject to their privacy policies. We encourage you to ask questions
-        before you disclose your personal information to others.
-      </p>
+      <p>Except as otherwise expressly included in this Privacy Policy, this document only addresses the use and disclosure of information we collect from you. To the extent that you disclose your information to other parties through the Site (including our payment processors), different rules may apply to their use, collection and disclosure of the personal information you disclose to them. Since we do not control the information use, collection or disclosure policies of third parties, you are subject to their privacy policies. We encourage you to ask questions before you disclose your personal information to others.</p>
 
       <H as="h2">International Transfers of Personal Information</H>
 
-      <p>
-        We store personal information about Users within Canada, the United
-        States and in other countries and territories. To facilitate our
-        operations, we may transfer and access such personal information from
-        around the world, including from other countries in which we have
-        operations. Therefore, if you are a resident of the European Economic
-        Area (the “EEA”), your personal information may be processed outside of
-        the EEA and in countries which are not subject to an adequacy decision
-        by the European Commission and which may not provide for the same level
-        of data protection as the EEA.
-      </p>
+      <p>We store personal information about Users within Canada, the United States and in other countries and territories. To facilitate our operations, we may transfer and access such personal information from around the world, including from other countries in which we have operations. Therefore, if you are a resident of the European Economic Area (the “EEA”), your personal information may be processed outside of the EEA and in countries which are not subject to an adequacy decision by the European Commission and which may not provide for the same level of data protection as the EEA.</p>
 
-      <p>
-        In this event, we will ensure that the recipient of your personal
-        information offers an adequate level of protection, for instance by
-        entering into standard contractual clauses for the transfer of data as
-        approved by the European Commission (Art. 46 GDPR), or we will ask you
-        for your prior consent to such international data transfers.
-      </p>
+      <p>In this event, we will ensure that the recipient of your personal information offers an adequate level of protection, for instance by entering into standard contractual clauses for the transfer of data as approved by the European Commission (Art. 46 GDPR), or we will ask you for your prior consent to such international data transfers.</p>
 
-      <p>
-        We have implemented safeguards to ensure an adequate level of data
-        protection where your personal information is transferred to countries
-        outside the EEA. For further details on the transfer of your personal
-        information, including a list of the organizations with whom said
-        information may be shared, please reach out to our team at the email
-        address provided below.
-      </p>
+      <p>We have implemented safeguards to ensure an adequate level of data protection where your personal information is transferred to countries outside the EEA. For further details on the transfer of your personal information, including a list of the organizations with whom said information may be shared, please reach out to our team at the email address provided below.</p>
 
       <H as="h2">Residents of the EEA and the State of California</H>
 
-      <p>
-        If you are a resident of the EEA or the State of California you may have
-        additional rights relating to the privacy of your information. We comply
-        with the privacy laws applicable to residents of the EEA and the State
-        of California. Please contact us at the email address provided below if
-        you have questions about your rights or wish to exercise them.
-      </p>
+      <p>If you are a resident of the EEA or the State of California you may have additional rights relating to the privacy of your information. We comply with the privacy laws applicable to residents of the EEA and the State of California. Please contact us at the email address provided below if you have questions about your rights or wish to exercise them.</p>
 
       <H as="h2">Questions?</H>
 
       <p>
-        It is our goal to make our privacy practices easy to understand. If you
-        have questions, concerns or if you would like more detailed information,
-        please email our privacy officer at{' '}
-        <a href="mailto:hey@wesbos.com">hey@wesbos.com</a>
+        It is our goal to make our privacy practices easy to understand. If you have questions, concerns or if you would like more detailed information, please email our privacy officer at <a href="mailto:hey@wesbos.com">hey@wesbos.com</a>
       </p>
     </>
   );
